Disable address validation if Smarty API erroring
If Smarty is returning errors (timeout, unavailable, bad requests, etc)
for any reason, aggressively switch to just accepting any provided
address. This flags in the database if the address was pre-validated or
not so that we can retroactively address any submissions while the API
was down.

Keeps timeout and retry configuration in the environment so we can
more easily adjust them if needed.